alterações em repositorys

+++ vehicleos/src/main/java/com/prova/Repositories/VehicleRepository.java
@@ -1,12 +1,11 @@
 package com.prova.repositories;
 
-import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.prova.domains.Vehicle;
 
 @Repository
-public interface VehicleRepository extends JpaRepository<Vehicle,long>{
+public interface VehicleRepository extends JpaRepository<Vehicle, Integer>{
 
 }
